requirements.txt
numpy>=1.24
glfw>=2.5.9
PyOpenGL>=3.1.6
Pillow>=10.0.0